requirements.txt
alembic==1.16.4
Authlib==1.6.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.34.131
botocore==1.34.131
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
extruct==0.18.0
feedparser==6.0.11
Flask==3.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.3
html5lib==1.1
html_text==0.7.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jstyleson==0.0.2
lxml==6.0.0
pandas==2.0.3
lxml_html_clean==0.4.2
Mako==1.3.10
MarkupSafe==3.0.2
mf2py==2.0.1
numpy>=1.24.0
pandas>=2.0.0
pycparser==2.22
pyparsing==3.2.3
pyRdfa3==3.6.4
python-dotenv==1.1.1
rdflib==7.1.4
recipe_scrapers==15.8.0
requests==2.32.4
scikit-learn>=1.3.0
sendgrid==6.11.0
sgmllib3k==1.0.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
stripe==12.3.0
torch>=2.0.0
transformers>=4.36.0
typing_extensions==4.14.1
urllib3==2.5.0
visitor==0.1.3
w3lib==2.3.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
Pillow==10.4.0
